Reorganizing packages to not load Express and work with non-Node and Browser loaders
