+++ package.json
@@ -22,7 +22,6 @@
   "bugs": {
     "url": "https://github.com/feathersjs/feathers/issues"
   },
-  "main": "lib/feathers",
   "directories": {
     "lib": "lib"
   },
@@ -48,6 +47,7 @@
   "dependencies": {
     "babel-polyfill": "^6.3.14",
     "debug": "^2.1.1",
+    "events": "^1.1.0",
     "express": "^4.12.3",
     "feathers-commons": "^0.5.0",
     "rubberduck": "^1.0.0",
@@ -60,8 +60,8 @@
     "babel-preset-es2015": "^6.3.13",
     "body-parser": "^1.13.2",
     "feathers-client": "^0.5.1",
-    "feathers-rest": "^1.0.0",
-    "feathers-socketio": "^1.0.0",
+    "feathers-rest": "^1.1.0",
+    "feathers-socketio": "^1.1.0",
     "istanbul": "^0.4.0",
     "jshint": "^2.6.3",
     "mocha": "^2.2.0",
@@ -70,8 +70,6 @@
     "socket.io-client": "^1.0.0"
   },
   "browser": {
-    "express": "./lib/client/express",
-    "./lib/feathers": "./lib/client/index",
-    "babel-polyfill": false
+    "./lib/index": "./lib/client/index"
   }
 }
