+++ src/client/index.js
@@ -1,6 +1,8 @@
 import feathers from '../feathers';
 import express from './express';
 
-export default function() {
-  return feathers(express());
+export default function createApplication(... args) {
+  return feathers(express(... args));
 }
+
+createApplication.version = require('../../package.json').version;
