+++ src/feathers.js
@@ -1,6 +1,3 @@
-if(!global._babelPolyfill) { require('babel-polyfill'); }
-
-import express from 'express';
 import Proto from 'uberproto';
 import Application from './application';
 
@@ -10,14 +7,8 @@ import Application from './application';
  * @return {Function}
  * @api public
  */
-export default function createApplication(app = express()) {
+export default function createApplication(app) {
   Proto.mixin(Application, app);
   app.init();
   return app;
 }
-
-// Framework version
-createApplication.version = require('../package.json').version;
-
-// Expose all express methods (like express.engine())
-Object.assign(createApplication, express);
